Add NoResolvablePureRefsChecker
Summary:
# Context
Having pure method/field refs left in the code is messy. Every pass that touches method/field refs has to rely on the Resolver to understand the state of that reference. More of then than not, similar logic is repeated to fight around it.

This change is part of an effort to tame the situation again around pure refs. We minimize the pure refs left in the code after the initial `ResolveRefsPass`. After that we want to make sure no unmanaged pure refs are introduced again in the code.

# This Change

- Add the property checker for method and field refs not being pure refs when resolvable
- The checker is currently only enabled for the instances of `ResolveRefsPass`
- We will roll it out more broadly in phases

Reviewed By: NTillmann

Differential Revision: D44936521

fbshipit-source-id: a2863413dbdca51e63cd53901c87202d9b715933