Adding option to prune uncallable instance methods
Summary:
In some cases, we don't need to keep the body of an uncallable method, but instead can make the method abstract, or remove it completely. We reuse the already existing logic from the RemoveUninstantiablesPass to perform the actual transformation.

This new behavior sits behind an option that's off by default, so this is a behavior-preserving change.

Reviewed By: thezhangwei

Differential Revision: D46615341

fbshipit-source-id: c9ba107db9916d7e78eb12d1ee414c5ea414cc00