Adding option to prune uncallable instance method bodies
Summary:
We can discard the body of instance methods that cannot be called, even though we might need to keep them for virtual scope order, or because they have keep rules.

To this end, we track what types are directly instantiable, and identify which instance methods are callable from those. We replace the body of all other (uncallable) instance methods with `throw null`, just like the RemoveUninstantiablesPass does (which will get disabled later in this stack).

The new behavior sits behind a new flag that is off by default, so this is a behavior-preserving change.

Reviewed By: thezhangwei

Differential Revision: D46583365

fbshipit-source-id: dcff4425db0adabaf2fcd929d6ecd28d22ef8658