For "dynamic references", only new-instance and const-class truely matter
Summary:
When looking at instructions to determine what class are "instantiable", the RemoveUninstantiablesPass historically only considered new-instance and const-class instructions. New-instance as it obviously creates an instance, and const-class as it may obtain a class object that can be used to create an instance via reflection.

However, DelInitPass, and since recently RMU, looked at all instruction with `has_type()`. RMU treats all those types as "dynamically referenced", which mimics the DelInitPass logic to determine for which classes to keep root instance members, as they may be used via reflection. However, the mere reference of a type via a check-cast or an instance-of instruction does in no way aid creating an instance via reflection. This seems to have been just an oversight in the DelInitPass. (new-array also refers to a type, albeit an array type, for which `type_class` always returns `nullptr`, so we can ignore that too.)

This is getting harmonized here by not considering types mentioned in check-cast or instance-of instructions anywhere when trying to figure out what types are instantiable.

Reviewed By: agampe

Differential Revision: D46844680

fbshipit-source-id: 511aed53485fb28decf0f79bc4bde5cba43cf95b