align interface instantiability with RemoveUninstantiablesPass
Summary:
Currently, RMU doesn't actually use the information about interfaces being marked as "instantiable" at all. So changing what gets marked, to align it with RemoveUninstantiablesPass, has no effect, and thus this is a behavior-preserving diff.

However, this information will be used in later diffs, when RMU will be taught to perform the functions of RemoveUninstantiablesPass.

Reviewed By: beicy

Differential Revision: D46568653

fbshipit-source-id: 97f9bf03f71f72cef7005c484fce6fd4f3a2d226