Fix forward_targets when move-result-*s are replaced
Summary:
This is a bug fix.

The issue is an interaction between 1) the constant-propagation transformation that rewrites move-result(-pseudo) instructions, and 2) the branch-target-forwarding.

The move-result-pseudo instruction might be in a different block than its associated primary instruction. The constant-propagation transformation generally uses the CFGMutator (yay), but it has a shortcoming, where it would put the replacement into the first block, even if the move-result instruction used to be in the second block. This is not generally a problem, but it is here, as the later running branch-target-forwarding transformation reuses the recorded abstract entry/exit states of blocks and their containing instructions, expecting them to be consistent, but now they are not. We work around that here by directly inserting in the right place, skipping the CFGMutator.

Reviewed By: wsanville

Differential Revision: D47083886

fbshipit-source-id: 6b0d7f8fc196e02ce7cfb700fe704a9d3d69ddd9