Update jobs (#795)
Summary:
Update the CircleCI jobs
* Add
   * Ubuntu 22.04
   * Debian Stable (currently 12)
   * Debian Unstable
* Move
   * Debian Stable -> Oldoldstable (currently 10)

The new containers are no longer CircleCI ones, as those have been deprecated in favor of a Ubuntu-only base build. Use the upstream `buildpack-deps` where necessary.

Those images, however, do not have a dedicated user and are quite minimal. Install `sudo` in the initial step to keep the rest of the steps the same. Add `unzip` which is required for testing.

Pull Request resolved: https://github.com/facebook/redex/pull/795

Differential Revision: D47160933

fbshipit-source-id: e1f6adafdc23f2791bd41d51c75e067b102f1273