Introducing NewObjectDomain / NewObjectAnalyzer
Summary:
This new domain is somewhat related to the existing `AbstractHeapPointer` + `ObjectDomain` / `ConstantArrayDomain`.

The new `NewObjectDomain` is very simple in a nature, just capturing the immutable aspects of any newly created object, whereas the other existing general object representation scheme is based on a heap indirection, which imposes a big overhead. Plus, unlike the simple new domain, the other heap tracking scheme allows for tracking of escaping, field values and array values --- however I must admit that I don't understand how the other heap tracking scheme would do that in a way that is always correct. So the `NewObjectDomain` does none of that.

The immutable aspects of any newly created object as its type, and in the case of an array, it's size.

Reviewed By: thezhangwei

Differential Revision: D46825206

fbshipit-source-id: d7fd931e11d082aa9627e641d26f46b1aa83540d