Revert D47298501: Multisect successfully blamed "D47298501: Back out "[redex][staging][stable][resolve-refs] Disable rtype specialization temporarily"" for test or build failures
Summary:
This diff is reverting D47298501
D47298501: Back out "[redex][staging][stable][resolve-refs] Disable rtype specialization temporarily" by thezhangwei has been identified to be causing the following test or build failures:

Tests affected:
- [automation_milan_appmanager_target30_mosnative_post_processed_rdk](https://www.internalfb.com/intern/test/844424977119621/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2397640
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: thezhangwei

Differential Revision: D47313247

fbshipit-source-id: 507d1d00fde09737455026f7f04dbe0641e15c91