Config option to preserve input dex files
Summary:
From what I can tell, the responsibilities of the python unpacker/repacker and redex-all binary are the following:
1) python unpacking logic creates an apk directory and a separate dex directory.
2) python layer also moves each dex file into a sub folder of the dex directory (i.e. classes.dex -> dex0/..., classes2.dex- > dex1/...)
3) redex-all binary outputs optimized dex files to the dex directory, which eventually will be packed back up along with the files in the apk directory.

For the use case of something like informational-only passes, or a config which is only meant to do dead resource elimination (like the one supplied) we may want an option to ensure that dex files are not changed unintentionally. It is not immediately clear to me if round tripping to our IR and back will always produce the same code, not ever change debug info, etc etc, so I am adding an option to this.

This change effectively turns step #2 above into a copy, tells redex-all binary to not write files, and asserts the file sizes have not been changed.

Reviewed By: agampe

Differential Revision: D47166165

fbshipit-source-id: 18cacd7e735da8d542d3efd7872677a348aed0d1