"Fix" `vencode`
Summary:
Increase buffer size. Use the redzone correctly - the buffer needs to be the full size.

A real fix should change the `encode` methods to use a `vector` or `deque` or similar for easy access as well as growing storage. Or an inspection of all callsites shows that pushing the caller's `vector` through may be acceptable.

Reviewed By: thezhangwei

Differential Revision: D47955944

fbshipit-source-id: 22580d0332907fd758b23c5b068f123daa3fa539