Move all header files into a `sparta` directory
Summary:
All sparta header files are currently in the root of the include directory.
It is usually recommended to have all the header files of a library under a single directory (e.g, `boost/`, `fmt/`, etc.).
This avoids name conflicts in the root.

Let's move all header files under a `sparta` directory.

Reviewed By: arnaudvenet

Differential Revision: D47903172

fbshipit-source-id: 5542009aaa72f57b527b4b5daeddeee525275f7e