add order_interdex option
Summary:
To enable experimentation to determine whether *ordering* matters. (The way the new option is implemented would still allow all other passes to make decisions based on the provided interdex classes, e.g. class merging and outlining.)
The new option is *on* by default, which corresponds to the current behavior. Thus, this is a behavior preserving change.

Reviewed By: beicy

Differential Revision: D47939851

fbshipit-source-id: 86f60103007326b5301065f1c5b4d70c10db22e7