Add a trait for patricia tree key encoding
Summary:
# Problem
Right now, we can only use unsigned integer types and pointers as the key to a patricia tree.
Actually, one could implement `sparta::pt_util::Codec<Key>`, but that interface is low level, one needs to reimplement the `decode` functions. It is also under `pt_util` and mostly undocumented, which means it's probably not meant to be overriden.

# Solution
This diff adds a proper trait for patricia tree keys, called `PatriciaTreeKeyTrait`, which defines the unsigned integer type to be used to encode the key. It can be easily extended by users of the library to use their own keys.

Reviewed By: arnaudvenet

Differential Revision: D48041843

fbshipit-source-id: ee205aad4881dd027edab50425b3f51dea8b2b26