ResFileManipulator
Summary:
While working on a new pass, it occurred to me that performing some types of edits in place from the libresource Visitor APIs is not very nice. For example, changing the size of a node in place makes things go haywire.

To make some jobs more convenient, define a helper class which will own copies of the data, put it in the right place when generating a final version of the file and stamp the size in the file header. Use it like `ButtonFiddler` in the test case (which overwrites the node with a new one with only 2 attributes) or as in D47616756.

Reviewed By: NTillmann

Differential Revision: D47741588

fbshipit-source-id: 05f2ff455e00655836b248f3586fb786cb7e5a88