Track in which classes implementation demands get fulfilled
Summary:
The method override graph now compute if methods from super types fulfill implementation demands from interfaces in subtypes.
This is exposed via a new optional filtering parameter to get_overriding_methods.

In later diffs, this makes it easier to do filtering of overriding methods based on the resolved (super) method of a particular invocation instruction with more specific type information encoded in the method ref.
For now, this new information isn't used yet, so this is a behavior-preserving change.

Reviewed By: agampe

Differential Revision: D48288172

fbshipit-source-id: 91c40f39032b0e2552ce62f91fe378f00b1e7ca9