Simplify moves before calling SwitchEquivFinder
Summary:
Perform a simplification step to make the existing logic in SwitchEquivFinder find more things. This step is optional and does not have to be comprehensive; it just makes the finder more likely to find applicable cases.

Test case is added to show why. Test is derived from a real world example in which a series of if checks against a const-class, after several optimization passes, contains a class checked from multiple blocks (facilitated by a move-object instruction). Before this diff, a move-object instruction would terminate the found cases, but with this diff more cases can be found.

Reviewed By: agampe

Differential Revision: D48360438

fbshipit-source-id: 758520bdd3e4ad7b720c7b025d9d5134305c865c