Break up code for easier reg-alloc
Summary:
Some enum code we encounter has a large live set of registers, which is detrimental to compilation speed (especially register allocation).

We can reorder the code and stay observationally equivalent (in non-OOM situations).

This also has the benefit of eliding many `move`s in the code and leads to a modest size improvement.

Reviewed By: thezhangwei

Differential Revision: D48279448

fbshipit-source-id: 6b39868300e423e94a7a6458e607737883cb9391