Skip abstract overridden method when merging true virtuals
Summary:
When merging true virtual methods, we sometimes have to emit an `invoke-super` call as a fallback for the mergeables do not override the base implementation of that method. Conceptually, we can skip that when every mergeable overrides that base.

However, in practice, to make things simple, we always emit the `invoke-super` call in the default block of the merged dispatch whenever we can find the overridden method. This logic could trigger a "risky pattern" when the overridden method is actually an abstract method. `invoke-super` on an abstract is invalid at Java source level. Although, it's not clear to me how it is enforced by the runtime.

Note that the redundant `invoke-super`s are most likely to be removed by subsequent passes so that they don't end up in our final release build. But to be more correct, we can skip emitting the call if the overridden method is abstract.

Reviewed By: wsanville

Differential Revision: D48419716

fbshipit-source-id: 3d660890c7578f8da2843127cf40f78985869d7e