use flags to indicate if a class rename status.
Summary: InitialRenameClassesPass (will be on top of this stack) needs to know which classes must be renamed, and which classes can not be renamed. Those info is calculated during RenameClassesPassV2 evaluation and stored with RenameClassesPassV2. To avoid fetching info from another pass, this diff adds 2 flags for each DexClass. This diff is behavior-preserving

Reviewed By: thezhangwei

Differential Revision: D48406422

fbshipit-source-id: aac5329d53476e98bf2e1afaceb1bc0954d00873