Complain loudly if unpacker cannot unpack to case insensitive file system
Summary: There have been several of these type issues https://github.com/facebook/redex/issues/822 where the unpacker is given an .apk file with obfuscation already applied, and unpacking to a case insensitive filesystem results in things not lining up. Try to detect this with an error to make it more obvious.

Reviewed By: agampe

Differential Revision: D48488524

fbshipit-source-id: 68d6618e6cfde1d8e0a30b2c85e6b32340814ffe