Better logging
Summary:
Record the state of dexes and try to figure out new entries so that aborts are more actionable. This is likely not very cheap, but since it is not permanently on it should be OK.

NOTE: It is strange that for the test case `DexLimitsInfo` reports a field overflow when in fact the methods are overflowing. This should be investigated.

Reviewed By: thezhangwei

Differential Revision: D48634505

fbshipit-source-id: e104c37f340147c179cf1bb8030d0253aa4ed509