sharpen overrides analysis for true virtuals
Summary:
Resolving a method refs "loses precision", as the resolved method may be in a super or interface type. When then looking at all overrides of the resolved method, we'll see many methods that cannot be possible overrides of the original (unresolved) method refs, which in fact can give us additional type information.

In this diff, we filter the obtained overridden methods accordingly.

The improved precision is used to 1) inform the inliner about unique callees for what looks like a true virtual invocation, 2) inform the inliner's cost function.

Reviewed By: beicy

Differential Revision: D48243456

fbshipit-source-id: b69dc443e84b3d7579e8962d8342dd0ea386f5f4