Back out "Use `nullptr` in ../fbandroid/native/redex/libresource/TypeWrappers.cpp" and others
Summary:
This backs out D48542268, D48541613, and D48542269 because r-barnes said that using the -Wno-zero-as-null-pointer-constant flag for the library should be fine to not block the high level goal. We did that in D48573515.

For context, this is an import of Android source code that we have in our project, and if you really want to change the code that is fine but we would like to do it in a way that applies it as a patch when importing a newer version of the code (there is a script for this). In any case, since we already discussed not changing the code at all, I am going to revert these.

Reviewed By: agampe

Differential Revision: D48762963

fbshipit-source-id: cd19dfdc08e1bcd6ad160ff27662fa99428fbb05