bug fix - don't remove only implementation methods in (instantiable) non-abstract classes
Summary:
We want to make sure that we don't remove the only implementation of inherited abstract methods in (instantiable) non-abstract classes.

This may not technically necessary, as verification at runtime is typically delayed to invocation time, and there should never be an invocation to an un-overridden abstract method, so that we should never run into issues anyway, but it's a nice property to preserve for sanity.

There were a few things that needed changing:
- For TAARMU (and in later diffs also regular RMU) we need to be careful to not remove "zombie" methods: Some directly instantiable classes may have vmethods that were not marked. Simply removing those methods might leave behind the class with unimplemented inherited abstract methods. We find if that's the case, d pick the first non-abstract override to add as an additional root.
- Similarly, when we `sweep_uncallable_virtual_methods`, we must not create new zombies.

Reviewed By: thezhangwei

Differential Revision: D48095563

fbshipit-source-id: 3f7e13ea747984a76db23d0b86a595781cfbe6b2