bug fix - classes that don't implement all inherited abstract methods should be marked abstract
Summary:
Classes that are not identified as instantiable need to be made abstract, as we may remove/ implementations/overrides from it.

This may not technically necessary, as verification at runtime is typically delayed to invocation time, and there should never be an invocation to an un-overridden abstract method, so that we should never run into issues anyway, but it's a nice property to preserve for sanity.

Reviewed By: thezhangwei

Differential Revision: D48096518

fbshipit-source-id: 24f0f0c778b2c88ebb26325b6dfea6913fa11bfc