new pseudo instruction: "unreachable"
Summary:
This new instruction must always come in the form
```
unreachable v0
throw v0
```
where `unreachable` can be thought of producing a throwable, which then must be immediately thrown.

One may ask why this complication, and not just have `unreachable` terminate the control-flow by itself. I tried that earlier, but that requires so many more changes everywhere, not just where control-flow termination matters, but also for proper monitor construction balancing with finally blocks. The design implemented in this diff avoids making all those changes (now), while achieving much of the goal.

What is the goal? Being able to distinguish a `throw null` in the code that might be reachable from (now) a `throw unreachable` which can be trusted to be unreachable due to earlier analysis.

This difference should help debugging, and matters to realize the change in the next diff.

This new instructions comes with 1) a lowering pass (just replacing the new instruction with a `const 0`), 2) a new property, and 3) a property checker.

Reviewed By: thezhangwei

Differential Revision: D48179328

fbshipit-source-id: 52e7bcb5d2242505d6ee878a10cd9b4c8ac59d2a