Move cls_to_interdex_group state to ConfigFiles
Summary:
To infer the interdex grouping of our mergeables, we first need to build a mapping from class to its betamap group id for classes listed in the betamap. We call that `cls_to_interdex_group`. The existing handling of building such map uses static state and is really dated.

This change moves the ownership of `cls_to_interdex_group` to ConfigFiles. That gives us a global "singleton" state that's nicely encapsulated in ConfigFiles. So that Class Merging doesn't need to worry about its detail and can just acquire it when needed lazily.

Reviewed By: NTillmann

Differential Revision: D48812676

fbshipit-source-id: 58aa1bb6f2aa7b45cd1fbc69749e43597ebf8027