InsertOnlyConcurrentMap
Summary:
Our ConcurrentMap never exposed the address of a stored value (except via explicitly documented unsafe and iterator methods).
The rationale was that this is "unsafe", as values might be concurrently mutated, or even erased. (There's a comment that also talks about addressed becoming invalid during internal hashtable resizing, but that's not actually the case.)

To address this, while still providing maximal safety in the face of concurrency, we here introduce a new "insert-only" concurrent map, whose elements can never be (safely) updated or erased. This in turn allows safely obtaining the address of a stored value in a concurrent scenario, avoiding copies (or other indirections). Later diffs will use this new functionality to reduce overhead.

Given a `InsertOnlyConcurrentMap<Key, Value>`, there are two new key functions (and overloads):
```
  const Value* get(const Key& key) const;

  template <typename ValueEqual = std::equal_to<Value>, typename... Args>
  std::pair<const Value*, bool> get_or_emplace_and_assert_equal(
      const Key& key, Args&&... args);
```
The `get` does the obvious lookup, returning `nullptr` when there is no value present. `get_or_emplace_and_assert_equal` looks up a value at a key; if the value is present, it asserts that the existing value is `==` to the value constructed with the given arguments; if the value is not present, it inserts a value constructed with the given arguments; finally, it returns the address of the value stored in the collection, and whether the provided value was emplaced.

Reviewed By: jimmycFB

Differential Revision: D49343487

fbshipit-source-id: 84efe307e1188156facc198ac851057f1f7d7ac8