resolve_invoke_method
Summary:
There are some invoke-virtual instructions that reference methods whose definition is actually in interface. Thus, for an invoke-virtual instruction, one should first try resolving it with `MethodSearch::Virtual`, and then `MethodSearch::InterfaceVirtual`. We used to have a bunch of places all over Redex where we do this little dance. This diffs consolidates all such dances with a new convenient helper function (which we should probably use in a few more places where we currently give up in such scenarios.)

This is a behavior-preserving change.

Reviewed By: thezhangwei

Differential Revision: D49440605

fbshipit-source-id: 2f77d1e6a4a5d22336a86911d309d7c4fcd0ba05