Implement a flat map following the patricia tree map interface
Summary:
This diff implements a map based on a sorted vector, following the interface of patricia tree maps.
This is very efficient for small maps thanks to CPU caches.
Following the interface of patricia tree maps will make it easy to use this to implement a partition or environment abstract domain. See follow-up diffs.

Reviewed By: arnaudvenet

Differential Revision: D49327533

fbshipit-source-id: 6014f747140dcf440c2f92457540e8e6040886c8