Implement `HashMap` as the base container for hashed abstract partition and environment
Summary:
`HashedAbstractPartition` and `HashedAbstractEnvironment` share a lot of similarities:
* They both have a `get` method that returns the default value
* They both use a "union" and "intersection"

This diff extracts a lot of the shared logic in a `HashMap` base class, which is generic hash map container using the same interface as `PatriciaTreeMap`.
This is then used to implement the `HashedAbstractPartition` and `HashedAbstractEnvironment` abstract domains, hence deduplicating their logic.

Reviewed By: arnaudvenet

Differential Revision: D49382373

fbshipit-source-id: f9047f116db747ed39aea261d0db213377d449f3