Fix NoSpuriousGetClass
Summary:
The code transformation is to avoid removing code that may be required for program semantics. Passes that might remove it should be annotated as requiring the transformation. Add RequiresAndPreserves for this.

NOTE: Current list of these was found by looking for the shrinker or known inlining passes (that do not turn off the shrinker). This list may be incomplete.

The property is also much better modeled as a negative property. It is automatically preserved and should not be in the final state.

Last but not least fix to preserve props where necessary.

Reviewed By: NTillmann

Differential Revision: D49714620

fbshipit-source-id: b8d628aa23f12184af2f3a7011048489cfbc691b