Implement a generalized hash map on top of patricia trees
Summary:
This diff implements a generalized hash map on top of patricia trees.
This allows to use patricia trees and their fast merging operators for any hashable key type - since `PatriciaTreeMap` only allows integer-type keys.
This works by using the hash of the key as the key in the patricia tree, and using a flat map as the value, to handle potential collisions.
Note that using the hash of the key is a common technique, for instance `immer` uses it to index elements into its B-tree like structure.
Also note that performance may vary a lot depending of your keys and the hash function. I recommend benchmarking before using this structure.

Reviewed By: arnaudvenet

Differential Revision: D49697555

fbshipit-source-id: 8aeb42eccb49ed5204a078583ea513c96c9b54f1