Implement a generic set abstract domain on top of a given set
Summary:
This implements a similar design as D50054646 but for sets.
Users can provide set implementations by implementing `AbstractSet`.
Then, a new class `SetAbstractDomain` can turn it into a powerset with a top and bottom element.
For now, this is only used to implement `PatriciaTreeSetAbstractDomain`, but we will use it for hash set and sparse set in a follow-up.

Reviewed By: arnaudvenet

Differential Revision: D50054946

fbshipit-source-id: 756a5b5a3b66b7525e8e06b6d5b0046daab8d07a