Add a base class for abstract maps
Summary:
We now have multiple implementations of "map" containers.
We want to create a generic adapter that gives a partition abstract domain and environment abstract domain for any map.
First, let's create a base class for abstract maps that enforces a common API.

Note that map APIs can have slight differences if the map is a mutable structure (hash map, flat map) or an immutable structure (patricia tree).
This works by enforcing the map to declare it's mutability and using `if constexpr` to check for the different interfaces.

The alternative would be to enforce a common interface, but that is a costly refactoring.

Reviewed By: arnaudvenet

Differential Revision: D50019979

fbshipit-source-id: 0a22e23a85111f2e3bfd7152b5c605ecbf49fe09