Rearrange Shape type naming scheme
Summary:
- Old naming scheme: `LPCreatorCreatorShape15S0000000_I1_9`
<root_type_tag>'Shape'<count>'S'<shape_layout_tag>'_I'<interdex_group_id>'_'<subgroup_idx>
There are two counts (`count` and `subgroup_idx`) that are inferred based on input order. They are not tied to the classes merged into the shape.
- New naming scheme: `LPCreatorCreatorShape_I1_S0000000_15_9`
<root_type_tag>'Shape_I'<interdex_group_id>'_S'<shape_layout_tag>'_'<count>'_'<subgroup_idx>
We move interdex grouping id to be before the shape layout. So that the alphanumerical order of the shape names is not sorted by the count in the old scheme.
- This change does not remove the counts yet. It rearranges the naming schema to be closer to the stable one.

Reviewed By: ssj933

Differential Revision: D49963774

fbshipit-source-id: dcbd56efe95dbc6550650f511162943bb724d81b