is_array_clone
Summary:
The `clone` method on any array does not have side-effects, but may read heap locations (the elements of the array it clones). It doesn't doesn't escape or return any parameters; it returns a new array.

Here, we encode these properties as appropriate summaries (instead of assuming the worst, which is the default).

Reviewed By: agampe

Differential Revision: D49939726

fbshipit-source-id: 490077ab5bf6c1cc4e8c0a9f701012756d298d6c