Revert D50081664: Multisect successfully blamed "D50081664: [llvm15] Migrate "fbandroid" from LLVM-12 to LLVM-15" for test or build failures
Summary:
This diff is reverting D50081664
D50081664: [llvm15] Migrate "fbandroid" from LLVM-12 to LLVM-15 by meyering has been identified to be causing the following test or build failures:

Tests affected:
- [fbsource//fbandroid/javatests/com/facebook/samples/assistant/playground/context:test - testQuiet (com.facebook.samples.assistant.playground.context.LoudnessJNITest)](https://www.internalfb.com/intern/test/844425028584910/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3330550
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: sheepsword

Differential Revision: D50400699

fbshipit-source-id: ad76412ddfdafc472b79d8be9ce4f9e358955164