Revert D50283299: Multisect successfully blamed "D50283282: [redex][string-int-def] Promote D49170241 and D50432485 to redex-staging and redex-stable" for test or build failures
Summary:
This diff is reverting D50283299
D50283282: [redex][string-int-def] Promote D49170241 and D50432485 to redex-staging and redex-stable by itang00 has been identified to be causing the following test or build failures:

Tests affected:
- [automation_milan_installer_target30_rdk](https://www.internalfb.com/intern/test/844424973383264/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3376238
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: itang00

Differential Revision: D50569026

fbshipit-source-id: 39393b1f992a9ee183899156b4fbb33708512c5a