InsertOnly*
Summary:
We should prefer using an InsertOnlyConcurrentMap (or an AtomicMap, which is coming in a later diff) when possibly:
InsertOnlyConcurrentMap more clearly conveys the possible intent of anonly map whose elements cannot be mutated, and it allows safely reading values without requiring copying them under a lock.

This is a behavior-preserving change.

Reviewed By: beicy

Differential Revision: D50439279

fbshipit-source-id: 93de65ae853f7193ee8e9b184ed79692df8cf1b9