Accessors on Visitor super class for headers, ids
Summary:
The chunk of the file that `visit_attribute_ids()` dispatches callbacks for is optional. However, the way it was implemented leaves a bit of ambiguity in that a subclass has no way of differentiating the scenario of "it was not present" vs. "it was present and has no ids".

That's getting changed here so that the callback will always happen if the chunk is present, and if it has no ids an explicit nullptr will be given so that if any caller makes a mistake it will be easily noticed (rather than just giving a pointer off the end of the data structure).

This diff also adds some simple accessors to the super class visitor for convenience, to access the position of this chunk (and others).

Reviewed By: NTillmann

Differential Revision: D50521165

fbshipit-source-id: 543ecf9186f1a162f07f6c23e4a1828c7dca954e