Make creation/destruction explicit
Summary:
This is to avoid undefined ordering between the destruction of the ThreadPool, and users of the thread-pool, as such the destructor of ConcurrentX datatypes.

As this change makes the existance of the threadpool optional, we need to adapt the PriorityThreadpool code to also work without a (general) threadpool (again).

Reviewed By: ssj933

Differential Revision: D50819078

fbshipit-source-id: 78e07a88a053c571b3984dbccefcd0fbe8e726b6