faster
Summary:
We used to do a lot of work before even scheduling the first task; much of that can be avoided. The main thing that's removed here is the `std::stable_sort` over all methods on the main thread.

We also switch over the handling of the wait-count to use lock-free operations.

This is a behavior-preserving change.

Reviewed By: agampe

Differential Revision: D50476083

fbshipit-source-id: 222b6a49669fdd9eebf1fdabb65f8b2794e691a4