Expanding XML editor to add attributes
Summary:
Expand the XML editor to add attributes. This includes adding an attribute under a new node. The editor will insert the attribute under the first node with a matching `tag_name`. If the attribute already exists, the XML editor will change the attribute's value to the one specified in the arguments.

How to:
To use the xml_editor to add an attribute to the AndroidManifest.xml file, do
`xml_editor AndroidManifest.xml <tag_name> <attribute name> <new attribute value in string/hex num> <attribute id> <attribute type> <attribute namespace> <node name>`

The last two arguments can be omitted if the attribute to insert has no namespace or if the attribute will not be inserted under a new node. If the namespace is provided, it must already exist in the file.

Example:
`xml_editor AndroidManifest.xml  manifest newAttribute 0x700 0x01010011 0x11 android newNode`
The above will add a new attribute named `newAttribute` under a newly created node called `newNode`. This new node will be under the `manifest` node, and the attribute will have the type `TYPE_INT_HEX` according to https://fburl.com/code/3xp4wgbd with a value of 0x700 and an id of 0x01010011.

If the attribute to insert does not contain an attribute id, substitute the attribute id argument with `0` instead.

More examples can be found in xml_editor_test.sh

Reviewed By: wsanville

Differential Revision: D47246963

fbshipit-source-id: d700b2efcdc2c4330501b5bc5075a251a9ea4a62