SynthAccessorPatcher
Summary:
# Context:

The frontend compiler could generate a synthetic accessor to facilitate accesses to private members of the enclosing scope from an inner class. In that case, the synth-accessor can invoke a callee with annotated parameters. However, the synth-accessor usually do not have the matching annotation replicated from its callee. So the invocation inside of it is a violation to our Typedef checker. Previously we have to exclude cases like that. But that also introduces small holes in our value safety enforcement.

# Changes:

- Introduce a `SynthAccessorPatcher` step in the `TypedefAnnoCheckerPass` that identifies all synth-accessors invoking an annotated callee
- The patcher replicates the Typedef annotation from the callee to the synth-accessor, so the annotation flow is complete

Reviewed By: itang00

Differential Revision: D51082376

fbshipit-source-id: 1d5ce4cc2ad3c2bf1497af6d161880be4960c7b4