ArtProfileWriter to add classes of hot methods.
Summary:
https://www.internalfb.com/code/fbsource/[178198e42623]/fbandroid/native/baseline_profiles/manual_profile.py?lines=47-48
this line of code in the baseline profile generator makes it clear that in manual profile (which includes profile from redex), a class has to be its own line in order to be picked into baseline profile, and eventually hopefully in app image.
Remember, a method line starts with its flag like 'H', "S', 'P', never 'L'.

So this change makes sure class of any hot method (and also startup-only method that will be included into profile) is included into baseline profile.

Reviewed By: NTillmann

Differential Revision: D51322912

fbshipit-source-id: 679ef61fdf41f67fc2c1cfa063e71b07f2608edb