lock-free string pool creation
Summary:
This is achieved by allowing to waste a few string ids during races.
This doesn't matter in the end, as we do a final round of deterministically ordering the string ids anyway, at which times any holes in the string ids get removed.

To minimize how many string ids we might waste initially, we cache the class/method string id lookup, and process positions in batches.

This is a behavior-preserving change.

Reviewed By: agampe

Differential Revision: D51519943

fbshipit-source-id: e5fe71d724a72ce05a92857c49737b1f893f5b71