Move static asserts in separate classes to enable default move constructors
Summary:
# Context
This is a follow-up of D39355689.
For short, having explicit destructors disables the generation of move constructors, which means we might be making copies when calling `std::move(x)` on those instances.

# Solution
Again, see D39355689.
We can move the static asserts in a destructor of an empty class, with the main class inheriting from it.

Reviewed By: agampe, arnaudvenet

Differential Revision: D51585268

fbshipit-source-id: d0d1f54c8086a49811f0a164fdcb20b88132b9d2