Add GlobalTypeAnalyzer::get_replayable_local_analysis() for local analysis replay
Summary:
# Context:

The `LocalTypeAnalyzer` by default includes `ClinitFieldAnalyzer` and `CtorFieldAnalyzer` that operate on a field type environment to model the heap when collecting type info for sfields and ifields. This design serves the WholeProgramState aggregation really well to not fall back to the previous global state too quickly. In the end, we aggregate all the field writes in the WholeProgramState, so everything is good.

However, after running the global type analysis, when replaying individual local analysis, for each individual field read, we should go straight to the global state. We should in this case, ignore the partial local info, and always go to the global one for completeness.

# This Change:

In this change, we split up the `LocalTypeAnalyzer` into two. The existing one operating on local field type environment remains the same, and serves WholeProgramState collection in the same way. We add a new one for replay purpose only after the global type analysis. The new replayable local analysis does not feed local field info into the register type environment. It always rely on the WholeProgramState for that.

Reviewed By: NTillmann

Differential Revision: D51539668

fbshipit-source-id: f765f0b912b286fe604baef95fd6417c5d4799c9