Skip ClinitFieldAnalyzer For WholeProgramState Aggregation
Summary: We use `ClinitFieldAnalyzer` to infer the type info of sfields the FinalInline style. However, the additional logic that come with it doesn't seem to produce a ton of value. There is also a potential risk of not evaluating things in the correct class initialization dependency order, like what we ran into with FinalInline. Therefore, here I propose to retire this small feature of the analysis.

Reviewed By: ssj933

Differential Revision: D51595560

fbshipit-source-id: c202198945f4e2aa9ccfe96ec0338b1e595e8fda