Remove non_returning_dependencies metric
Summary:
The metric is not determistic, but doesn't affect the final result.

Background:
When coming across an interface/virtual method invocation, RMU might suspend scanning of the control-flow graph any further, until the invoked method or one of its overrides has been found to contain a return statement. When suspending scanning, dependencies on all relevant invoked methods and their overrides are registered. There can be race conditions between whether a method is first found to be returning, or if first an invocation to a method is encountered when it is still unknown whether the method can returning. Depending on these races, dependencies can be registered or not. The final result is independent of these races. However, the metric on whether dependencies were registered is not.

Reviewed By: ssj933

Differential Revision: D51720362

fbshipit-source-id: f7b71f11207c4b77bdfbb3d4430a331bfcc5d2af