Support checking synthetic stubs for Kotlin default argument
Summary:
# Context:

Kotlin support default arguments. To implement that in bytecode, `kotlinc` synthesizes a stub with a `$default` suffix in its name to supply the default argument value to the actual method. The synthesized stub does not carry the same param annotation from its callee. Out of the box, our Typedef annotation checker doesn't properly cover this case.

# Solution

The `SynthAccessorPatcher` can patch the param annotation for the `$default` stub similar to what it does to the synthetic accessor. Here we simply remove the guard that excludes the `$default` stubs from the checker logic.

We also added new tests to cover both the calls to the `$default` stubs as well as the calls from the `$default` stubs.

Reviewed By: itang00

Differential Revision: D51904548

fbshipit-source-id: cd3cecb390dceca1bee986d89840374675a817f0