determinism
Summary: Tracking down the non-determinism we recently saw, I found that it may involve joining "singleton objects" and "objects with immutable attr"; this happens in the inliner when it does interprocedural constant-propagation to find constant arguments. These particular domains are not disjoint, and we've been fighting with non-determinism issues with them from the beginning. Reading through the current join and meet implementation, I discovered that the current implementation is clearly imprecise, losing too much information. This is getting fixed here.

Reviewed By: thezhangwei

Differential Revision: D52136179

fbshipit-source-id: e07e96fc1e785190817267b825ddd419a3e3c858