Replace the guts of SwitchMethodPartitioning with SwitchEquivFinder
Summary:
SwitchEquivFinder is a more capable version of representing a switch (or series of if statements) that came after SwitchMethodPartitioning. SwitchEquivFinder is more robust in handling some of the choices d8 could make when representing a switch that was written in the source code i.e. with multiple compares that jump to the same successor.

There isn't really a need to keep around the two different implementations, so this diff keeps the SwitchMethodPartitioning class, comments surrounding it and its simplified representation and fills that in with the details from SwitchEquivFinder. Overall, I view SwitchMethodPartitioning as a special case of what the finder does.
- Special case for methods that always throw (i.e. some empty codegen method).
- `verify_default_case_throws` is useful to make sure that SwitchEquivFinder doesn't stop midway through a method and bail out, assuming something it can't understand is the default case.

Reviewed By: thezhangwei

Differential Revision: D52064112

fbshipit-source-id: 0ecc0a908811f8013d133ec3ec54b70da868e1a1