Move SwitchMethodPartitioning
Summary:
SwitchMethodPartitioning was moved from its optimization to open source since it was touted as a general purpose utility. SwitchEquivFinder came later and is better at handling such jobs, and in the dependent diffs SwitchMethodPartitioning became just a holder for SwitchEquivFinder results that is catered to its caller.

I see no value in having SwitchMethodPartitioning as a general purpose thing, move it to its use and further specialize it to do that task. Anything that wanted to use its functionality should be using SwitchEquivFinder anyways.

Reviewed By: thezhangwei

Differential Revision: D52213332

fbshipit-source-id: d198909c168431c2803ea89f3e163223a95b1b56