bug fix
Summary: When the enum analysis could determine that a register holds a value of a transformable enum type that provides a custom toString() method, but then `Enum.toString()` is called on that value (as opposed to `CustomEnumType.toString()`), then illegal code would get generated (an invoke-static to the virtual method `Enum.toString()`). This is getting fixed here.

Reviewed By: beicy

Differential Revision: D52420175

fbshipit-source-id: d7291a3d57cc44284c9cc8736a872004bdea05a0