"internal"
Summary:
There are currently two different places where newly generated methods are adorned with "$oea$": 1) for reduced variants of root methods, and 2) for expanded methods where an object argument is replaced with a sequence field values. This is confusing. This diff makes it so that 1) gets a different name variation to make it easier to distinguish when debugging. Eventually, all those "internal" methods will get removed anyway.

This is a behavior-preserving change.

Reviewed By: beicy

Differential Revision: D52410659

fbshipit-source-id: 3a595b5861d7da63dc11de22e0285d4ec76e41fb