Change generating build (#833)
Summary:
Pull Request resolved: https://github.com/facebook/redex/pull/833

The cache was misconfigured. A "save" was triggered for Ubuntu 20.04 - which does not build Boost from source. So the cache was empty.

Change the generating build to be Ubuntu 18.04. While this is technically outdated, it is the build we originally pushed. Eventually this should move to the Debian 10 one (oldoldstable).

`setup_oss_toolchain.sh` was using a temp directory for the tools. Under CircleCI use the project directory so the cache works.

Add a cache for protobuf.

Remove the python3.6 code.

Increase the boost cache key version.

Reviewed By: jimmycFB

Differential Revision: D52492772

fbshipit-source-id: bc5191abd5e9955b76a983afb4a52db4c923e5d1