ResolveRefsPass test case for external methods
Summary:
Corresponding test case for the fix D52631931. The test case fails without D52631931, passes with it. Much of the boilerplate here is to set up an integ test case that works with android.jar and satisfies ExternalRefsManglingPass so that `m_refine_to_external` is true (instr test would have been "easier" but I wasn't able to get the right input opcodes to create the problem).

If/when ResolveRefsPass gets upgraded to change the invoke opcode if we want to swap in the interface toString(), we can just update this test case accordingly.

Reviewed By: NTillmann

Differential Revision: D52641858

fbshipit-source-id: d3d29a8e6e9698d6df3d5ea70f560e2e8385d7c7